custom (and also buggy) base64 processing code replaced with the core functions
git-svn-id: http://svn.miranda-ng.org/main/trunk@4987 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/CryptoPP/src/base16.cpp
@@ -18,7 +18,7 @@ char *base16encode(const char *inBuffer, int count) {
 }
 
 
-char *base16decode(const char *inBuffer, size_t *count) {
+char *base16decode(const char *inBuffer, unsigned int *count) {
 
 	char *outBuffer = (char *) malloc(*count);
 	BYTE *outBufferPtr = (BYTE *) outBuffer;
@@ -56,7 +56,7 @@ char *base16decode(const char *inBuffer, size_t *count) {
 
 char *base16decode(const char *inBuffer)
 {
-	size_t count = strlen(inBuffer);
+	unsigned count = (unsigned)strlen(inBuffer);
 	return base16decode(inBuffer, &count);
 }
